requirements.txt
anthropic>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
click>=8.0.0
